数据库存储short 有问题，因此修改为int 增加TimePatterDesc类 TimeBaseServiceImpl类型修改后，修改月份。取int字节的0，1 位 去除去调试的部分

+++ app/src/main/java/cn/com/aiton/domain/GbtTimeBase.java
@@ -12,7 +12,7 @@ public class GbtTimeBase {
 
 	private int timeBaseId;
 
-	private short month;
+	private int month;
 
 	private int week;
 
@@ -44,11 +44,11 @@ public class GbtTimeBase {
         this.timeBaseId = timeBaseId;
     }
 
-    public short getMonth() {
+    public int getMonth() {
         return month;
     }
 
-    public void setMonth(short month) {
+    public void setMonth(int month) {
         this.month = month;
     }
